use Test::Perl::Critic unconditionally
Stop skipping the test if for some reason Test::Perl::Critic is not
available: tests are useless if they are skipped for bad reasons.
And in the previous commit we ensured that Test::Perl::Critic will
appear in `dzil listdeps --develop`, so there is really no reason for
that to happen.